pyproject.toml
[tool.poetry]
name = "projects-visualization"
version = "0.1.0"
description = ""
authors = ["alejolop1 <clalejo.0803@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
mysql-connector-python = "^9.2.0"
pygwalker = "^0.4.9.13"
plotly = "^5.24.1"
sqlalchemy = "^2.0.37"
streamlit-autorefresh = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
